+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - "data" to job description create
 - Custom fields to Job Descriptions
 - Add custom data to job description search results
+- Add international_country_code to phone number details
 
 ## [5.3.1] - 2023-03-30
 ### Changed